+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 
 ==========
 
+## Ver. 1.0.3
+- Set sync false to pipeline
+- Add video passthrough
+
+## Ver. 1.0.2
+- Check if duration is not -1
+
 ## Ver. 1.0.1
 - Update dependencies